pipeline/requirements.txt
feedparser==6.0.10
google-api-core==1.27.0
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-language==2.0.0
google-cloud-translate==3.1.0
googleapis-common-protos==1.53.0
pandas==1.2.4
tqdm==4.64.1
torch==1.13.0
transformers==4.13.0